True Pitch Implementation *WIP*
Takes approximately 30 seconds to properly initialize, as it starts reading nose low (10+ degrees) and slowly centers out. Not totally sure why this is, but it could have a negative effect on True Booster if you mount soon after turning the board on.

This issue was present previously on 5.3 too, though I'm not sure to this degree. This probably should be resolved and not swept under the rug.