InputTilt Updates
Setting "Remote Type" to an option beyond "NONE" will cause the Balance App to crash when engaging (or if the rider writes this change while riding) if they are not on the custom firmware that supports the needed function.

While it is simply resolved with a reboot and switching back to NONE, there does technically pose the safety risk of someone (for some reason) switching to it while riding. Preferably, we would either be able to check firmware version or the set "App to Use" before attempting the function call.