Disabled Booster Speed Stiffness
A sudden jerk would be felt when transitioning between no Speed Stiffness and applying Speed Stiffness, as well as going from accel to braking due to the doubled factor on braking.

If included in the future, this should be properly ramped in, rather than suddenly adding / removing current from Applied Booster Current.