Revert broken SoftStart Fix
Was fix for SoftStart not working when engaging from Nose Down position. While it worked, it also caused Rate P to be limited in normal riding, beyond the startup phase. Really not sure why this is or why the loop is active beyond startup at all, but that can be tackled another day.