Migrate build system to Makefiles and generate confxml via build system
Adds a hierarchy of Makefiles to replace the build_res script. This
allows to build packages independently and properly specify their target
dependencies.

Also removes the built .vescpkgs from git, these are now always built
using the full VESC Tool invocation.

The VESC Tool default command has been changed to vesc_tool (assuming
most users will have it in PATH is better than a hard-coded path that
may not apply to a lot of users), and it can be overriden with any make
invocation like so:

make VESC_TOOL=/usr/local/bin/vesc_tool

For the float package, makes the version extractor from settings.xml
more robust, finding the version in the 10 lines that follow the
APPCONF_FLOAT_VERSION id occurrence in the xml, not just the 8th one
(still not great, but better).

Generate confxml sources via the build system:

Removes the confxml.{c,h} from git and adds targets to automatically
generate them at build time. Removes the error-prone manual step of
having to generate them manually and commit to git.

The binary data commited to git are hard to verify for correctness and
is in essence a duplication of information.