Don't modify FS state with simplestart/posi, evaluate when needed
The simplestart/posi features shouldn't override the footpad sensor
state to FS_BOTH, as that should represent the true value of the
sensors. Instead, add a function that evaluates whether the sensors are
engaged and which takes simplestart and posi into account.

One of the practical advantages is the user is now able to see the true
state of the sensors when only one side is activated, as opposed to
always seeing both sides are.