Remove torqueresponse from current limiting
Assuming torqueresponse setpoint offset of 5 degrees, the limit would be
increased by 0.06 * 5 = 0.3A. This is a very insignificant amount.

Also, the limit would be increased above the actual motor current limit,
which would only make sense if the limit was lower than what is set in
VESC, and that's not the case.