Replace old State and SetpointAdjustmentType enums with new ones
The State was an agglomeration of semantically unrelated state
components. It had trouble representing combinations of the state
components in a single flat list, required workarounds and introduced
unnecessary corner cases. Instead, use a clean RunState enum that mainly
just distinguishes between READY and RUNNING, and add a handful of other
attributes to a new State struct to represent the overall state in a
concise way.

The pushback state is represented by SetpointAdjustmentType (which was
just reordered, renamed and renumbered), darkride is an extra flag,
flywheel is a special mode of operation (along with handtest), and
wheelslip is also a flag. The various FAULT_ states are moved out to a
new StopCondition enum.